% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhtmlPalmTrees.R
\name{PalmTrees}
\alias{PalmTrees}
\title{Create a PalmTree plot}
\usage{
PalmTrees(data = NULL, weights = NULL, hover.color = "#eeeeee",
  row.names = NULL, row.font.size = 11, row.font.family = "sans-serif",
  row.font.color = "#000000", row.heading = NULL,
  row.heading.font.size = 12, row.heading.font.family = "sans-serif",
  row.heading.font.color = "#000000", col.names = NULL,
  col.font.size = 11, col.font.family = "sans-serif",
  frond.color.unselected = "#cccccc", frond.color.this = "#000000",
  frond.color.that = "#cccccc", col.font.color = "#000000",
  col.font.color.unselected = "#aaaaaa", col.heading = NULL,
  col.heading.font.size = 12, col.heading.font.family = "sans-serif",
  col.heading.font.color = "#000000", tooltips = TRUE,
  tooltips.font.size = 11, tooltips.font.family = "sans-serif",
  tooltips.heading.font.size = 12,
  tooltips.heading.font.family = "sans-serif", y.show = TRUE,
  y.lab = NULL, y.digits = 1, y.font.size = 11,
  y.font.family = "sans-serif", y.font.color = "#000000",
  y.lab.font.size = 12, y.lab.font.family = "sans-serif",
  y.lab.font.color = "#000000", prefix = NULL, suffix = NULL,
  digits = 1, colors = NULL, order = "descending",
  legend.background.color = "#ffffff", legend.maxProportion = 0.25,
  legend.border.color = "#000000", width = NULL, height = NULL)
}
\arguments{
\item{data}{numeric matrix or data frame}

\item{weights}{numeric vector specifying weights. Length must equal to \code{ncol(data)}}

\item{hover.color}{color of cells that are hilighted by mouse interaction}

\item{row.names}{character vector specifying x ticks. Default value are obtained from \code{rownames(data)}}

\item{row.font.size}{Default x axis font size = 11}

\item{row.font.family}{x axis font family = "sans-serif"}

\item{row.font.color}{font color for x axis text}

\item{row.heading}{character, used as x label}

\item{row.heading.font.size}{Default row heading font size = 12}

\item{row.heading.font.family}{Default row heading font family = "sans-serif"}

\item{row.heading.font.color}{font color for x axis title text}

\item{col.names}{character vector specifying the legend. Default value obtained from \code{colnames(data)}}

\item{col.font.size}{Default column font size = 11}

\item{col.font.family}{Default column font family = "sans-serif"}

\item{frond.color.unselected}{color used on unselected fronds on the main plot}

\item{frond.color.this}{color used in legend to show frond placement in each frond legend row}

\item{frond.color.that}{color used in legend for other fronds in frond legend row}

\item{col.font.color}{legend font color}

\item{col.font.color.unselected}{legend fond color to indicate inactive selections (rows and orders)}

\item{col.heading}{character vector specifying the heading}

\item{col.heading.font.size}{Default column heading font size = 12}

\item{col.heading.font.family}{Default column heading font family = "sans-serif"}

\item{col.heading.font.color}{legend title font color}

\item{tooltips}{logical, whether to attach tooltips on mouseover, default true}

\item{tooltips.font.size}{Default tooltip font size = 11}

\item{tooltips.font.family}{Default tooltip font family = "sans-serif"}

\item{tooltips.heading.font.size}{Default tooltip heading font size = 12}

\item{tooltips.heading.font.family}{Default tooltip heading font family = "sans-serif"}

\item{y.show}{show the y axis}

\item{y.lab}{y axis labels}

\item{y.digits}{integer to control the number of decimal places of the y axis}

\item{y.font.size}{Default y axis font size = 11}

\item{y.font.family}{Default y axis font family = "sans-serif"}

\item{y.font.color}{y axis font color}

\item{y.lab.font.size}{Default y axis label font size = 12}

\item{y.lab.font.family}{Default y axis label font family = "sans-serif"}

\item{y.lab.font.color}{y axis labels font color}

\item{prefix}{prefix of numbers in the tooltips. If suffix is not provided, prefix will take suffix' place on the y axis}

\item{suffix}{suffix of numbers in the tooltips}

\item{digits}{integer to control the number of decimal places in the tooltips}

\item{colors}{colors of the fronds. D3 colors will be used if no values are provided}

\item{order}{= c("original", "alphabetical", "ascending", "descending") specifies the column order with default = "descending"}

\item{legend.background.color}{the canvas background color}

\item{legend.maxProportion}{Max proportion of legend width to plot width. Default = 0.25}

\item{legend.border.color}{legend border color}

\item{height}{}
}
\description{
Create a PalmTree plot
}
\details{
Frond size is calculated by mapping \code{data} to a linear scale \code{[min_width, max_width]} so that
\code{min(data)} takes \code{min_width} and \code{max(data)} takes \code{max_width}.
\code{min_width} is currently set at 8 pixcels and \code{max_width} is calculated using the available space between x ticks.
}
\examples{

data("CSDperceptions", package = "rhtmlPalmTrees")

weights = rep(1,ncol(CSDperceptions))
prefix = ""
suffix = "\%"
rhtmlPalmTrees::PalmTrees(data = CSDperceptions,
                         weights = weights,
                         row.names = rownames(CSDperceptions),
                         row.heading = names(dimnames(CSDperceptions))[1],
                         col.names = colnames(CSDperceptions),
                         col.heading = names(dimnames(CSDperceptions))[2],
                         prefix = prefix,
                         suffix = suffix,
                         tooltips = TRUE,
                         colors = qColors)



}
