% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhtmlPalmTrees.R
\name{PalmTrees}
\alias{PalmTrees}
\title{Create a PalmTree plot}
\usage{
PalmTrees(data = NULL, weights = NULL, row.names = NULL,
  row.heading = NULL, col.names = NULL, col.heading = NULL,
  tooltips = TRUE, prefix = NULL, suffix = NULL,
  column.as.heights = NULL, y.lab = NULL, y.prefix = NULL,
  y.suffix = NULL, colors = NULL, order = "descending", width = NULL,
  height = NULL)
}
\arguments{
\item{data}{numeric matrix or data frame}

\item{weights}{numeric vector specifying weights}

\item{row.names}{character vector specifying x ticks}

\item{row.heading}{character, used as x label}

\item{col.names}{character vector specifying the legend}

\item{col.heading}{character vector specifying the heading}

\item{tooltips}{logical, whether to attach tooltips on mouseover, default true.}

\item{prefix}{prefix of numbers in the tooltips. If column.as.heights is NULL (default), the y axis will also have the same prefix. If suffix is not provided, prefix will take suffix' place on the y axis.}

\item{suffix}{suffix of numbers in the tooltips. If column.as.heights is NULL (default), the y axis will also have the same suffix.}

\item{column.as.heights}{Integer to specify which column can be used as tree heights.}

\item{y.lab}{y axis label}

\item{y.prefix}{prefix of y axis ticks. This argument is ignored when column.as.heights is NULL (default), in which case prefix is used.}

\item{y.suffix}{suffix of y axis ticks. This argument is ignored when column.as.heights is NULL (default), in which case suffix is used.}

\item{colors}{colors of the leaves  (should be optional but now it must be provided)}

\item{order}{= c("original", "alphabetical", "ascending", "descending") specifies the column order with default = "descending".}

\item{width}{}

\item{height}{}
}
\description{
Create a PalmTree plot
}
\examples{
data("CSDperceptions", package = "rhtmlPalmTrees")

weights = rep(1,ncol(CSDperceptions))
prefix = ""
suffix = "\%"
rhtmlPalmTrees::PalmTrees(data = CSDperceptions,
                         weights = weights,
                         row.names = rownames(CSDperceptions),
                         row.heading = names(dimnames(CSDperceptions))[1],
                         col.names = colnames(CSDperceptions),
                         col.heading = names(dimnames(CSDperceptions))[2],
                         prefix = prefix,
                         suffix = suffix,
                         tooltips = TRUE,
                         colors = qColors)
}

